requirements.txt
openai>=1.40.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
transformers>=4.46.0
accelerate>=0.34.0
safetensors>=0.4.5
huggingface_hub>=0.24.0
